// src/app/tui/models/tarjetacirculacion.ts
export class TarjetaCirculacion {
    idTARJETACIRCULACION: number;
    nroExpediente:string;
    nroBoletaPago: string;
    fechaPago: string;
    nroSocio: number;
    montoTC: number;
    montoIncVeh: number;
    idEMPRESA: number;
    fechaExpedicion: string;
    fechaVencimiento: string;
    idVEHICULO: number;
    estado: boolean;

}